refactor: features/games/ (GamesPage, dialogs, detail/, expansions/, characters/)
