+++ src/components/StatsPage.tsx
@@ -3,7 +3,7 @@ import { useSearchParams } from 'react-router-dom';
 import { ArrowLeft, Users, TrendUp } from '@phosphor-icons/react';
 import { useQuery } from '@tanstack/react-query';
 import { playerApi } from '@/services/api/playerApi';
-import { gameApi } from '@/services/api/gameApi';
+import { gameApi } from '@/features/games/gameApi';
 import { queryKeys } from '@/shared/services/api/queryKeys';
 import { useNavigationAdapter } from '@/shared/hooks/useNavigationAdapter';
 import { useLabels } from '@/shared/hooks/useLabels';
