+++ src/components/dialogs/index.ts
@@ -2,12 +2,3 @@
 export { AddPlayerDialog } from './AddPlayerDialog';
 export { EditPlayerDialog } from './EditPlayerDialog';
 export { DeletePlayerDialog } from './DeletePlayerDialog';
-
-// Game dialogs
-export { default as AddGameDialog } from './AddGameDialog';
-export { default as EditGameDialog } from './EditGameDialog';
-export { default as DeleteGameDialog } from './DeleteGameDialog';
-
-// Expansion and Character dialogs
-export { AddExpansionDialog, EditExpansionDialog, DeleteExpansionDialog } from './ExpansionDialogs';
-export { AddCharacterDialog, EditCharacterDialog, DeleteCharacterDialog } from './CharacterDialogs';
