+++ src/features/games/characters/GameCharactersView.tsx
@@ -3,8 +3,8 @@ import { Button } from '@/shared/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/shared/components/ui/card';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/shared/components/ui/tooltip';
 import { ArrowLeft, Plus, PencilSimple, Trash, User, Lightning } from '@phosphor-icons/react';
-import { useGameCharacters, UseGameCharactersProps } from '@/hooks/games/useGameCharacters';
-import { AddCharacterDialog, EditCharacterDialog, DeleteCharacterDialog } from '@/components/dialogs';
+import { useGameCharacters, UseGameCharactersProps } from './useGameCharacters';
+import { AddCharacterDialog, EditCharacterDialog, DeleteCharacterDialog } from './dialogs/CharacterDialogs';
 
 export default function GameCharactersView(props: UseGameCharactersProps & { darkMode: boolean }) {
   const {
