+++ src/features/games/detail/GameDetailPage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { useGameDetail, UseGameDetailProps } from '@/hooks/games/useGameDetail';
-import GameDetailView from '@/views/games/GameDetailView';
+import { useGameDetail, UseGameDetailProps } from './useGameDetail';
+import GameDetailView from './GameDetailView';
 
 interface GameDetailPageProps extends UseGameDetailProps {
   currentView: string;
