+++ src/features/games/detail/GamePageRoute.tsx
@@ -1,11 +1,11 @@
 import { useParams } from 'react-router-dom';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
-import { gameApi } from '@/services/api/gameApi';
+import { gameApi } from '@/features/games/gameApi';
 import { queryKeys } from '@/shared/services/api/queryKeys';
 import { useNavigationAdapter } from '@/shared/hooks/useNavigationAdapter';
 import GameDetailPage from './GameDetailPage';
-import GameExpansionsPage from './GameExpansionsPage';
-import GameCharactersPage from './GameCharactersPage';
+import GameExpansionsPage from '../expansions/GameExpansionsPage';
+import GameCharactersPage from '../characters/GameCharactersPage';
 import type { GameExpansion, GameCharacter } from '@/types';
 
 type SubRoute = 'detail' | 'expansions' | 'characters';
