+++ src/features/games/dialogs/index.ts
@@ -1,0 +1,3 @@
+export { default as AddGameDialog } from './AddGameDialog';
+export { default as EditGameDialog } from './EditGameDialog';
+export { default as DeleteGameDialog } from './DeleteGameDialog';
