+++ src/features/games/expansions/GameExpansionsView.tsx
@@ -3,8 +3,8 @@ import { Button } from '@/shared/components/ui/button';
 import { Card, CardContent, CardHeader, CardTitle } from '@/shared/components/ui/card';
 import { Tooltip, TooltipContent, TooltipTrigger } from '@/shared/components/ui/tooltip';
 import { ArrowLeft, Plus, PencilSimple, Trash, Calendar } from '@phosphor-icons/react';
-import { useGameExpansions, UseGameExpansionsProps } from '@/hooks/games/useGameExpansions';
-import { AddExpansionDialog, EditExpansionDialog, DeleteExpansionDialog } from '@/components/dialogs';
+import { useGameExpansions, UseGameExpansionsProps } from './useGameExpansions';
+import { AddExpansionDialog, EditExpansionDialog, DeleteExpansionDialog } from './dialogs/ExpansionDialogs';
 import { GameExpansion } from '@/types';
 import { getContentClass } from '@/shared/utils/gameHelpers';
