+++ src/hooks/useNewGamePage.ts
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useQuery, useMutation, useQueryClient } from '@tanstack/react-query';
 import { toast } from 'sonner';
-import { gameApi } from '../services/api/gameApi';
+import { gameApi } from '@/features/games/gameApi';
 import { playerApi } from '../services/api/playerApi';
 import { playApi } from '../services/api/playApi';
 import { queryKeys } from '@/shared/services/api/queryKeys';
